# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "InteractiveGenerateRequest": {
        "description": "互动内容生成请求：备课意图（可透传上次备课的意图对象）。",
        "example": {
          "intent": {
            "grade": "初二",
            "interactive": "课堂小游戏：判断哪些是一次函数",
            "topic": "一次函数"
          }
        },
        "properties": {
          "intent": {
            "additionalProperties": true,
            "default": {},
            "title": "Intent",
            "type": "object"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "title": "InteractiveGenerateRequest",
        "type": "object"
      },
      "InteractiveGenerateResponse": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "html": {
            "title": "Html",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "html",
          "filename"
        ],
        "title": "InteractiveGenerateResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "多模态 AI 备课智能体后端。\n\nOpenAPI 是本服务接口文档的**唯一事实源**：每个端点都带 `summary`、描述、\n请求/响应示例、错误码与分组 tag；前端 TypeScript 类型由本 schema 生成，禁止手抄。\n* **界面**：Swagger UI `/docs`，ReDoc `/redoc`。\n* **分组**：系统 / 备课会话 / 知识库 / 知识图谱 / 生成物 / 冲突审核 / 题库。\n* **单用户**：固定 `user_id=\"default\"`，本服务不含登录与多租户。\n* **启动清扫**：进程重启时，上一轮被中断的「处理中」资料不会有后台任务再推进，\n  服务启动时统一标记为「失败」，避免前端无限轮询（教师可按原文重新上传）。\n* **能力切换**：对话模型、向量化、语音转写、PDF 解析、网络搜索均按配置选择。\n  云端能力未配置时相关端点返回 `503`（`code: provider_not_configured`，`message` 面向教师，\n  指到设置页）——产品不返回任何「看起来像真结果的占位内容」。\n* **生成失败**：配好了但生成没成功时返回 `502`（`code: generation_failed`），\n  教师可重试——不落一份空课件 / 空教案冒充成功。\n* **版本策略**：只有破坏性变更才升 `/api/v2`。\n",
    "title": "EduMind API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/interactive/generate": {
      "post": {
        "description": "按意图生成一份**单文件 HTML5** 互动内容（小游戏 / 知识点动画），落盘后返回 HTML 正文与文件名。\n\n前端可用 `GET /api/v1/files/{filename}?inline=true` 在新标签页直接打开试用。\n模型返回的不是完整单文件 HTML 时，本接口以 502 报错而不落盘半成品。\n\n**带上 `session_id` 时本份互动内容落一条新版本**（版本号在会话内单调递增，旧的互动内容版本保持可回看、可下载，也可用版本下载端点直内联打开）；会话不存在返回 `404`。",
        "operationId": "generate_interactive_api_v1_interactive_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractiveGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "filename": "creative_3c2b1a09.html",
                  "html": "<!doctype html><html lang=zh><body>…</body></html>",
                  "session_id": "9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b",
                  "version": 1,
                  "version_id": "8f7e6d5c-4b3a-4291-9c8d-7e6f5a4b3c2d"
                },
                "schema": {
                  "$ref": "#/components/schemas/InteractiveGenerateResponse"
                }
              }
            },
            "description": "互动内容已生成并落盘"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "会话不存在: 9f1a2b3c-4d5e-4f60-8a7b-1c2d3e4f5a6b"
                }
              }
            },
            "description": "备课会话不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "messages"
                      ],
                      "msg": "Field required",
                      "type": "missing"
                    }
                  ]
                }
              }
            },
            "description": "请求校验失败:请求体、表单或路径字段缺失或类型不符"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            },
            "description": "未捕获的服务端错误:进程存活但本次请求失败,前端应提示重试。"
          },
          "502": {
            "content": {
              "application/json": {
                "example": {
                  "detail": "互动内容生成失败：模型未返回单文件 HTML，请重试"
                }
              }
            },
            "description": "互动内容生成失败：模型未返回单文件 HTML"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "code": "provider_not_configured",
                    "message": "还没有配置模型供应商：到「设置 → 添加供应商」选一家并粘贴 Key 后即可使用。"
                  }
                }
              }
            },
            "description": "所需的云端能力未配置：不返回演示用假结果。`detail.code` 固定为 `provider_not_configured`，`detail.message` 是面向教师的一句引导（去设置页添加供应商或补齐配置）。配置后即时生效，无需重启。"
          }
        },
        "summary": "生成互动内容",
        "tags": [
          "生成物"
        ]
      }
    },
...
  },
...
}